Get SubscriptionWorker working as a Sidekiq job
- Delegate work to Poll to test separately from Sidekiq
